Quickly add in darkside for SSE-based change notification
This is a waaay simpler approach. Yes, it means we have to maintain
connections instead of offloading that to Google and friends, but Safari
doesn't even support web push, and it's just problematic anyway. In
particular, with the old web push approach there's no great time to
let the server know that we're no longer interested in a channel, which
means we could receive changes when the browser is closed and be
required to show a notification, which would be super annoying.

Now, if the user goes away, they don't get the push. Simple.

This also means that, especially once I remove the server-side gapi
calls (because while that was cool in theory, the access tokens we could
get in the browser just don't last long enough to be useful) there's
really going to be very little actually stored in the Redis nodes (if
anything?) aside from whatever it needs for pubsub. So, it should be
very easy to scale out horizontally if we ever need to (hopefully not).