Begin building out services for new push protocol
After thinking about it for a while, handling push notifications for
changes from a 3rd party service is very tricky. Dropping web-push is
definitely the right move, because there's already quite a bit of
complexity here.

In particular, we definitely have to create the watches on the server
in order to prevent spoofing from malicious users—they wouldn't be able
to actually *modify* anything, and there are steps the server could take
to prevent them from tricking other clients into refreshing excessively,
but in order to avoid getting potentially hundreds of duplicate "file
changed" notifications for the same file (and accidentally DDOSing ourselves),
we would like to request just *one* person (or, if this is problematic,
just a few people) actually be in charge of watching the file. If we
trust all the clients when they say they have access to do this, some
bad actors could *claim* to be able to watch a file, then just sit on
their hands.

So, we try to create watches on the server, where we can know if they
actually have access or not.

Right now there's still an opportunity if a bad actor randomly gets
selected when the original watcher leaves (since we don't store the auth
and just ask that they send it back) but that seems unlikely. If we see
this happening a lot then we can request multiple people try to watch a
file, and there's de-duping logic in place during a race.