Relax requirement that the "watcher" of a sheet be online
Fixes #5
Refs #1

It really doesn't matter *who* created the watch, so long as *someone*
did. The watch can last for up to 24 hours on Google; we don't need
*nearly* that long, but we can use 4 hours as a comfortable duration for
a D&D session.

If the watch *does* expire and there are still people interested in the
sheet, they will be listening on a channel with the sheet's ID. Right
now, every such user is informed of the need to watch, and *all*
well-behaved clients will then respond with their attempt to create the
watch. This is fine for now, but future work will want to only ask a
random sampling of all listeners (see #1). I don't expect we'll ever
have that kind of user base, but Critical Role's Campaign 2 premier
apparently had something like 150K viewers; if such an event ran long,
we probably would prefer *not* to rapidly get 150K requests.