Refactor `sse` service to `channels`; add socket.io channel support
Refs #6

Sadly, now.sh doesn't play nicely with SSE, so as an alternative we have
to use socket.io. We don't have as much control over which members of
a "room" to send to (at least, not yet?) which is unfortunate, but
hopefully future work can port similar logic to our socket.io
implementation.

This is a good refactor to do anyway. The services shouldn't (and,
indeed, didn't) need to know about the SSE implementation stuff; that's
API-level.