Add socket.io channel support (#7)
* Rename watch.send -> watch.notifyChanged

* Refactor `sse` service to `channels`; add socket.io channel support

Refs #6

Sadly, now.sh doesn't play nicely with SSE, so as an alternative we have
to use socket.io. We don't have as much control over which members of
a "room" to send to (at least, not yet?) which is unfortunate, but
hopefully future work can port similar logic to our socket.io
implementation.

This is a good refactor to do anyway. The services shouldn't (and,
indeed, didn't) need to know about the SSE implementation stuff; that's
API-level.

* Refactor to handle all distributed events in-house

Closes #6

Much simpler. Much less load on Redis. We just use one channel, and
socket.io doesn't have to do... whatever it is socket.io-redis was
trying to do. I'm still not sure because it never actually worked
correctly.